requirements.txt
discord.py
aiosqlite
python-dotenv
google-generativeai
yt-dlp
